tests/wpt/web-platform-tests/dom/events/Event-dispatch-on-disabled-elements.html

<!doctype html>
<meta charset="utf8">
<meta name="timeout" content="long">
<title>Events must dispatch on disabled elements</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<body>
<script>
// HTML elements that can be disabled
const formElements = ["button", "fieldset", "input", "select", "textarea"];

test(() => {
  for (const localName of formElements) {
    const elem = document.createElement(localName);
    elem.disabled = true;
    // pass becomes true if the event is called and it's the right type.
    let pass = false;
    const listener = ({ type }) => {
      pass = type === "click";
    };
    elem.addEventListener("click", listener, { once: true });
    elem.dispatchEvent(new Event("click"));
    assert_true(
      pass,
      `Untrusted "click" Event didn't dispatch on ${elem.constructor.name}.`
    );
  }
}, "Can dispatch untrusted 'click' Events at disabled HTML elements.");

test(() => {
  for (const localName of formElements) {
    const elem = document.createElement(localName);
    elem.disabled = true;
    // pass becomes true if the event is called and it's the right type.
    let pass = false;
    const listener = ({ type }) => {
      pass = type === "pass";
    };
    elem.addEventListener("pass", listener, { once: true });
    elem.dispatchEvent(new Event("pass"));
    assert_true(
      pass,
      `Untrusted "pass" Event didn't dispatch on ${elem.constructor.name}`
    );
  }
}, "Can dispatch untrusted Events at disabled HTML elements.");

test(() => {
  for (const localName of formElements) {
    const elem = document.createElement(localName);
    elem.disabled = true;
    // pass becomes true if the event is called and it's the right type.
    let pass = false;
    const listener = ({ type }) => {
      pass = type === "custom-pass";
    };
    elem.addEventListener("custom-pass", listener, { once: true });
    elem.dispatchEvent(new CustomEvent("custom-pass"));
    assert_true(
      pass,
      `CustomEvent "custom-pass" didn't dispatch on ${elem.constructor.name}`
    );
  }
}, "Can dispatch CustomEvents at disabled HTML elements.");

test(() => {
  for (const localName of formElements) {
    const elem = document.createElement(localName);

    // Element is disabled... so this click() MUST NOT fire an event.
    elem.disabled = true;
    let pass = true;
    elem.onclick = e => {
      pass = false;
    };
    elem.click();
    assert_true(
      pass,
      `.click() must not dispatch "click" event on disabled ${
        elem.constructor.name
      }.`
    );

    // Element is (re)enabled... so this click() fires an event.
    elem.disabled = false;
    pass = false;
    elem.onclick = e => {
      pass = true;
    };
    elem.click();
    assert_true(
      pass,
      `.click() must dispatch "click" event on enabled ${
        elem.constructor.name
      }.`
    );
  }
}, "Calling click() on disabled elements must not dispatch events.");

promise_test(async () => {
  // Style sheet that controls transition.
  const style = document.createElement("style");
  style.innerText = `
    ${formElements.join(", ")} {
      opacity: 0.1;
      transition-property: opacity;
      transition-duration: .1s;
    }
    .transition {
      opacity: 1;
    }
  `;
  document.head.appendChild(style);

  // Triggers the transition in the element being tested.
  const transitionTrigger = document.createElement("button");
  transitionTrigger.innerText = "Trigger button";
  document.body.appendChild(transitionTrigger);

  // For each form element type, set up transition event handlers.
  for (const localName of formElements) {
    const elem = document.createElement(localName);
    elem.disabled = true;
    document.body.appendChild(elem);
    const transitionPromises = [
      "transitionrun",
      "transitionstart",
      "transitionend",
    ].map(eventType => {
      return new Promise(r => {
        const handlerName = `on${eventType}`;
        elem[handlerName] = ev => {
          elem[handlerName] = null;
          r();
        };
      });
    });

    // Trigger transitions specifically on this element
    // it requires a trusted event.
    transitionTrigger.onclick = () => {
      elem.classList.toggle("transition");
    };
    await test_driver.click(transitionTrigger);

    // All the events fire...
    await Promise.all(transitionPromises);
    elem.classList.remove("transition");

    // Let's now test the "transitioncancel" event.
    elem.ontransitionstart = () => {
      // Cancel the transition by hiding it.
      elem.style.display = "none";
      elem.classList.remove("transition");
    };

    // Trigger the transition again!
    const promiseToCancel = new Promise(r => {
      elem.ontransitioncancel = r;
    });
    await test_driver.click(transitionTrigger);
    await promiseToCancel;
    // And we are done with this element.
    elem.remove();
  }
  // And we are done with the test... clean up.
  transitionTrigger.remove();
  style.remove();
}, "CSS Transitions events fire on disabled form elements");

promise_test(async () => {
  const style = document.createElement("style");
  style.innerText = `
    .animate {
      animation: fade .1s 2;
    }
    @keyframes fade {
      0% { opacity: 1; }
      100% { opacity: 0.2; }
    }
  `;
  document.head.appendChild(style);
  // For each form element type, set up transition event handlers.
  for (const localName of formElements) {
    const elem = document.createElement(localName);
    document.body.appendChild(elem);
    elem.disabled = true;
    const transitionPromises = [
      "animationstart",
      "animationiteration",
      "animationend",
    ].map(eventType => {
      return new Promise(r => {
        elem.addEventListener(eventType, r, { once: true });
      });
    });
    elem.classList.add("animate");
    // All the events fire...
    await Promise.all(transitionPromises);
    elem.remove();
  }
  // And we are done with the test... clean up.
  style.remove();
}, "CSS Animation events fire on disabled form elements");

promise_test(async () => {
  const style = document.createElement("style");
  style.innerText = `
    .animate {
      animation: fade .1s 2;
    }
    @keyframes fade {
      0% { opacity: 1; }
      100% { opacity: 0.2; }
    }
  `;
  document.head.appendChild(style);
  // For each form element type, set up transition event handlers.
  for (const localName of formElements) {
    const elem = document.createElement(localName);
    document.body.appendChild(elem);
    elem.disabled = true;
    // Let's now test the "animationcancel" event
    const promiseToCancel = new Promise(r => {
      elem.addEventListener("animationcancel", r);
    });
    elem.addEventListener("animationstart", () => {
      // Cancel the animation by hiding it.
      elem.style.display = "none";
    });
    elem.classList.add("animate");
    // Trigger the animation again!
    await promiseToCancel;
    // And we are done with this element.
    elem.remove();
  }
  // And we are done with the test... clean up.
  style.remove();
}, "CSS Animation's cancel events fire on disabled form elements");

promise_test(async () => {
  for (const localName of formElements) {
    const elem = document.createElement(localName);
    elem.disabled = true;
    document.body.appendChild(elem);
    // Element is disabled, so clicking must not fire events
    let pass = true;
    elem.onclick = e => {
      pass = false;
    };
    // Disabled elements are not clickable.
    await test_driver.click(elem);
    assert_true(
      pass,
      `${elem.constructor.name} is disabled, so onclick must not fire.`
    );
    // Element is (re)enabled... so this click() will fire an event.
    pass = false;
    elem.disabled = false;
    elem.onclick = () => {
      pass = true;
    };
    await test_driver.click(elem);
    assert_true(
      pass,
      `${elem.constructor.name} is enabled, so onclick must fire.`
    );
    elem.remove();
  }
}, "Real clicks on disabled elements must not dispatch events.");
</script>
